+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
+## [1.3.5r3] - 2021-06-16
+
 ### Changed
   - participants.tsv table is managed using bidsMeta.BidsTable class